src/main/java/com/mangione/continuous/sampling/SampledObservationProvider.java
package com.mangione.continuous.sampling;

<<<<<<< HEAD
import com.mangione.continuous.observationproviders.ObservationProviderInterface;
import com.mangione.continuous.observationproviders.RandomGeneratorFactory;
import com.mangione.continuous.observations.ObservationInterface;

import javax.annotation.Nonnull;
import java.util.Iterator;
import java.util.Spliterator;
import java.util.function.Consumer;


public class SampledObservationProvider<S, T extends ObservationInterface<S>> implements ObservationProviderInterface<S, T> {
	private final double testSamplingPercent;
	private final ObservationProviderInterface<S, T> provider;
	private final RandomGeneratorFactory generatorFactory;
	private final long seed;
	private final boolean isTestSet;
	private final long numberOfObservations;


	@SuppressWarnings("WeakerAccess")
	public SampledObservationProvider(double testSamplingPercent, ObservationProviderInterface<S, T> provider,
			RandomGeneratorFactory generatorFactory, long seed, boolean isTestSet) {
		this(testSamplingPercent, provider, generatorFactory, seed, isTestSet, provider.getNumberOfLines());

	}

	public SampledObservationProvider(double testSamplingPercent, ObservationProviderInterface<S, T> provider,
	                                  RandomGeneratorFactory generatorFactory, long seed, boolean isTestSet, long numberOfObservations) {
		this.testSamplingPercent = testSamplingPercent;
		this.provider = provider;
		this.generatorFactory = generatorFactory;
		this.seed = seed;
		this.isTestSet = isTestSet;
		this.numberOfObservations = numberOfObservations;
=======
import java.util.Comparator;
import java.util.Iterator;
import java.util.List;
import java.util.function.Consumer;

import javax.annotation.Nonnull;

import com.mangione.continuous.observationproviders.ListObservationProvider;
import com.mangione.continuous.observationproviders.ObservationProviderInterface;
import com.mangione.continuous.observations.ObservationInterface;
import com.mangione.continuous.sampling.reservoir.SamplerFactoryInterface;
import com.mangione.continuous.sampling.reservoir.SamplerInterface;


public class SampledObservationProvider<S, T extends ObservationInterface<S>> implements ObservationProviderInterface<S, T> {
	private final ListObservationProvider<S, T> sampledObservations;
>>>>>>> 73d9563 (Migrated file changes from the source.)

	public SampledObservationProvider(ObservationProviderInterface<S, T> providerToSample, int sampleSize, int seed, SamplerFactoryInterface<T> factory) {
		SamplerInterface<T> sampler = factory.createSampler(providerToSample.iterator(), sampleSize, seed);
		sampledObservations = new ListObservationProvider<>(sampler.sample());
	}

	public SampledObservationProvider(ObservationProviderInterface<S, T> providerToSample, int sampleSize, int seed, SamplerFactoryInterface<T> factory, Comparator<T> sortFunction) {
		SamplerInterface<T> sampler = factory.createSampler(providerToSample.iterator(), sampleSize, seed);
		List<T> sample = sampler.sample();
		sample.sort(sortFunction);
		sampledObservations = new ListObservationProvider<>(sample);
	}

	@Override
	@Nonnull
	public Iterator<T> iterator() {
		return sampledObservations.iterator();
	}

	@Override
	public void forEach(Consumer<? super T> action) {
		Iterator<T> iterator = iterator();
		iterator.forEachRemaining(action);
	}
<<<<<<< HEAD

	@Override
	public Spliterator<T> spliterator() {
		throw new UnsupportedOperationException("Spliterator not supported.");
	}

	private class SampledObservationProviderIterator implements Iterator<T> {
		private final Iterator<T> iterator = provider.iterator();
		private final SamplingWithoutReplacement samplingWithoutReplacement;
		private boolean isCurrentSampleTest;
		private T nextObservation;


		private SampledObservationProviderIterator() {
			samplingWithoutReplacement =
					new SamplingWithoutReplacement(testSamplingPercent, numberOfObservations,
							generatorFactory.generate(seed));
			if (iterator.hasNext())
				getNextAndUpdateCurrentSampleTest();
		}

		@Override
		public boolean hasNext() {
			while (iterator.hasNext() && isCurrentSampleTest != isTestSet) {
				getNextAndUpdateCurrentSampleTest();
			}
			return iterator.hasNext();
		}

		@Override
		public T next() {
			while (iterator.hasNext() && isCurrentSampleTest != isTestSet) {
				getNextAndUpdateCurrentSampleTest();
			}

			T next = null;
			if (isCurrentSampleTest == isTestSet) {
				next = nextObservation;
				if (iterator.hasNext())
					getNextAndUpdateCurrentSampleTest();

			}
			return next;
		}

		@Override
		public void remove() {
			throw new UnsupportedOperationException("remove not supported.");

		}

		@Override
		public void forEachRemaining(Consumer<? super T> action) {
			while (hasNext()) {
				action.accept(next());
			}
		}

		private void getNextAndUpdateCurrentSampleTest() {
			isCurrentSampleTest = samplingWithoutReplacement.select();
			nextObservation = iterator.next();
		}
	}
=======
>>>>>>> 73d9563 (Migrated file changes from the source.)
}
